Const qualifiers and cleanup across codebase
Small API and style cleanups across native and managed code: bump GitHub Action versions and global .NET SDK; add missing const qualifiers and tighten parameter types in many C/C++ files; fix null-check/bracing and callback signatures for GTK/Win32 integrations; refactor Linux taskbar DBus variable names and error handling; minor Windows/WebView and notification API tweaks; simplify a few C# members (expression-bodied property, remove unused using, Enum.TryParse usage) and adjust tests. These changes improve consistency, reduce compiler warnings and tighten APIs.

+++ src/InfiniFrame.Shared/Utilities/MonitorsUtility.cs
@@ -62,7 +62,7 @@ internal static class MonitorsUtility {
             Rectangle intersection = Rectangle.Intersect(m.MonitorArea, windowBounds);
             long overlap = 0;
             if (intersection.Width > 0 && intersection.Height > 0) {
-                overlap = (long)intersection.Width * (long)intersection.Height;
+                overlap = intersection.Width * (long)intersection.Height;
             }
 
             // fraction of the *window* that lies on this monitor
